refactor: removed a lot of functions from mail client
