+++ internal/mail/mock_client.go
@@ -8,3 +8,4 @@ func (*TestClient) Send(ctx context.Context, user *User, body string) error {
 	logger.Info("Test sending mail", "user", user.Email)
 	return nil
 }
+
